Moved DatasetLocationIndexer into the dataset module
