+++ backend/dataall/modules/datasets/indexers/__init__.py
@@ -1,0 +1,1 @@
+"""Contains dataset related indexers for OpenSearch"""
